api dang ky mon that param is list

+++ src/main/java/com/qlsv5/service/impl/CommonServiceImpl.java
@@ -242,6 +242,7 @@ public class CommonServiceImpl implements CommonService {
             result.setId(UUID.randomUUID().toString().split("-")[0]);
             return chiTietLopTcRepository.save(result);
         }
+        /* fact : not use */
         else if(object instanceof DiemDto){
 
             /* TRUTH: DANG KY MON */
@@ -257,6 +258,31 @@ public class CommonServiceImpl implements CommonService {
 
             return diemRepository.save(result);
         }
+        else if(object instanceof DangKyMonDto){
+
+            /* TRUTH: DANG KY MON */
+            DangKyMonDto dangKyMonDto = modelMapper.map(object, DangKyMonDto.class);
+
+            List<DiemDto> listDiem = new ArrayList<>();
+            for(int i = 0 ; i < dangKyMonDto.getDiemDtoList().size() ; i++){
+                listDiem.add(dangKyMonDto.getDiemDtoList().get(i));
+            }
+
+            for(int i = 0 ; i < listDiem.size() ; i++){
+                DiemEntity result = new DiemEntity();
+                result = modelMapper.map(listDiem.get(i), DiemEntity.class);
+                result.setId(UUID.randomUUID().toString().split("-")[0]);
+
+                /*SET SOLUONGCON CUA LOPTINCHI*/
+                DsLopTcEntity dsLopTcEntity = dsLopTcRepository.getDsLopTcByMaLopTc(result.getMaLopTc());
+                dsLopTcEntity.setSoLuongCon(dsLopTcEntity.getSoLuongCon() - 1);
+
+                dsLopTcRepository.save(dsLopTcEntity);
+
+                diemRepository.save(result);
+            }
+            return null;
+        }
 
         return null;
     }
@@ -268,10 +294,22 @@ public class CommonServiceImpl implements CommonService {
 
         if(object instanceof KhoaDto){
             for (String item : lstId) {
+
                 int countMaKhoa = khoaRepository.countKhoaById(item);
+
                 if(countMaKhoa > 0){
-                    lstSuccess.add(item);
-                    khoaRepository.deleteById(item);
+                    /* important: validator */
+                    KhoaEntity khoaEntity = khoaRepository.findById(item).get();
+                    int countMaKhoaOnGiangVienTable = giangVienRepository.countGiangVienByMaKhoa(khoaEntity.getMaKhoa());
+                    int countMaKhoaOnLopTable = lopRepository.countLopByMaKhoa(khoaEntity.getMaKhoa());
+
+                    if(countMaKhoaOnLopTable > 0 || countMaKhoaOnGiangVienTable > 0){
+                        continue;
+                    }
+                    else {
+                        lstSuccess.add(item);
+                        khoaRepository.deleteById(item);
+                    }
                 }
             }
         }
@@ -279,8 +317,18 @@ public class CommonServiceImpl implements CommonService {
             for (String item : lstId) {
                 int countMaLop = lopRepository.countLopById(item);
                 if(countMaLop > 0){
-                    lstSuccess.add(item);
-                    lopRepository.deleteById(item);
+                    /* important: validator */
+                    LopEntity lopEntity = lopRepository.findById(item).get();
+                    int countMaLopOnSinhVienTable = sinhVienRepository.countSinhVienByMaLop(lopEntity.getMaLop());
+                    int countMaLopOnDsLopTcTable = dsLopTcRepository.countDsLopTcByMaLop(lopEntity.getMaLop());
+
+                    if(countMaLopOnSinhVienTable > 0 || countMaLopOnDsLopTcTable > 0){
+                        continue;
+                    }
+                    else {
+                        lstSuccess.add(item);
+                        lopRepository.deleteById(item);
+                    }
                 }
             }
         }
@@ -288,8 +336,19 @@ public class CommonServiceImpl implements CommonService {
             for (String item : lstId) {
                 int countMaSinhVien = sinhVienRepository.countSinhVienById(item);
                 if(countMaSinhVien > 0){
-                    lstSuccess.add(item);
-                    sinhVienRepository.deleteById(item);
+
+                    /* important: validator */
+                    SinhVienEntity sinhVienEntity = sinhVienRepository.findById(item).get();
+                    int countSinhVienOnDiemTable = diemRepository.countDiemByMaSv(sinhVienEntity.getMaSv());
+
+                    if(countSinhVienOnDiemTable > 0){
+                        continue;
+                    }
+                    else {
+                        lstSuccess.add(item);
+                        sinhVienRepository.deleteById(item);
+                    }
+
                 }
             }
         }
@@ -297,8 +356,19 @@ public class CommonServiceImpl implements CommonService {
             for (String item : lstId) {
                 int countMaGiangVien = giangVienRepository.countGiangVienById(item);
                 if(countMaGiangVien > 0){
-                    lstSuccess.add(item);
-                    giangVienRepository.deleteById(item);
+
+                    /* important: validator */
+                    GiangVienEntity giangVienEntity = giangVienRepository.findById(item).get();
+                    int countGiangVienOnDsLopTcTable = dsLopTcRepository.countDsLopTcByMaGv(giangVienEntity.getMaGv());
+
+                    if(countGiangVienOnDsLopTcTable > 0){
+                        continue;
+                    }
+                    else {
+                        lstSuccess.add(item);
+                        giangVienRepository.deleteById(item);
+                    }
+
                 }
             }
         }
@@ -306,8 +376,19 @@ public class CommonServiceImpl implements CommonService {
             for (String item : lstId) {
                 int countMaMonHoc = monHocRepository.countMonHocById(item);
                 if(countMaMonHoc > 0){
-                    lstSuccess.add(item);
-                    monHocRepository.deleteById(item);
+
+                    /* important: validator */
+                    MonHocEntity monHocEntity = monHocRepository.findById(item).get();
+                    int countGiangVienOnDsLopTcTable = dsLopTcRepository.countDsLopTcByMaMh(monHocEntity.getMaMh());
+
+                    if(countGiangVienOnDsLopTcTable > 0){
+                        continue;
+                    }
+                    else {
+                        lstSuccess.add(item);
+                        monHocRepository.deleteById(item);
+                    }
+
                 }
             }
         }
@@ -315,8 +396,29 @@ public class CommonServiceImpl implements CommonService {
             for (String item : lstId) {
                 int countMaDsLopTc = dsLopTcRepository.countDsLopTcById(item);
                 if(countMaDsLopTc > 0){
-                    lstSuccess.add(item);
-                    dsLopTcRepository.deleteById(item);
+
+                    /* important: validator */
+                    DsLopTcEntity dsLopTcEntity = dsLopTcRepository.findById(item).get();
+                    int countDslopTcOnDiemTable = diemRepository.countDiemByMaLopTc(dsLopTcEntity.getMaMh());
+                    int countDslopTcOnChiTietLopTcTable = chiTietLopTcRepository.countChiTietLopTcByMaLopTc(dsLopTcEntity.getMaMh());
+
+                    if(countDslopTcOnDiemTable > 0 || countDslopTcOnChiTietLopTcTable > 0){
+                        continue;
+                    }
+                    else {
+
+                        /*delete Loptc -> delte all ChiTietLopTc that relative with LopTc*/
+                        List<ChiTietLopTcEntity> chiTietLopTcEntityList = chiTietLopTcRepository.getListChiTietLopTcByMaLopTc(dsLopTcEntity.getMaLopTc());
+
+                        for (ChiTietLopTcEntity chiTietLopTcEntity: chiTietLopTcEntityList) {
+                            chiTietLopTcRepository.deleteById(chiTietLopTcEntity.getId());
+                        }
+                        /* end */
+
+                        lstSuccess.add(item);
+                        dsLopTcRepository.deleteById(item);
+                    }
+
                 }
             }
         }
